Authentication, Friends and small UI updates
Changed how emails are verified when logging in and signing up. This is now done using Firebase instead of client side.

Add friends input verification works and checks if the user exists on the platform. To be developed is the actual add friends functionality.

Smooth fade in and out for modals.
Background image for login and signup is slightly reduced in size for faster loading. Need to figure out a caching system for image and video assets which are not modified frequently